+++ SwiftyJSONAPI/JSONAPIDocument.swift
@@ -23,11 +23,11 @@ public class JSONAPIDocument: JSONPrinter {
     public convenience init(_ json: [String:AnyObject]) {
         self.init()
         for object in normalizeJSONAPIObjectToArray(json["data"]) {
-            data.append(JSONAPIResource(object))
+            data.append(JSONAPIResource(object, parentDocument: self, loaded: .NotLoaded))
         }
         
         for object in normalizeJSONAPIObjectToArray(json["included"]) {
-            included.append(JSONAPIResource(object))
+            included.append(JSONAPIResource(object,parentDocument: self))
         }
         
         if let strings = json["links"] as? [String:String] {
